requirements.txt
discord
python-dotenv
google-generativeai>=0.7.2